implemented HW time; meinberg time backend

+++ src/ullc.c
@@ -83,9 +83,9 @@ INA_API(ina_rc_t) ina_ullc_get_ring_info(const char *name, ina_ullc_rb_info_t *i
     }
 
     info->ring_version = rb->version;
-    info->num_write_op = rb->next_ptr;
+    info->num_write_op = (size_t)rb->next_ptr;
     info->last_writer = 0;
-    info->num_read_op = rb->cursor;
+    info->num_read_op = (size_t)rb->cursor;
     info->last_reader = 0;
     info->num_producers = rb->num_producers;
     info->num_producers_alive = 0;
